refactor: Extract memory pool to separate package
