+++ internal/extractor/extractor_test.go
@@ -13,6 +13,7 @@ import (
 	"go.mongodb.org/mongo-driver/mongo/options"
 
 	"mongo2dynamo/internal/common"
+	"mongo2dynamo/internal/pool"
 )
 
 // MockCollection is a mock implementation of Collection interface.
@@ -92,8 +93,8 @@ func TestMongoExtractor_Extract_EmptyCollection(t *testing.T) {
 	mockCursor.On("Err").Return(nil)
 	mockCollection.On("Find", mock.Anything, mock.Anything, mock.Anything).Return(mockCursor, nil)
 
-	docPool := common.NewDocumentPool()
-	chunkPool := common.NewChunkPool(2000)
+	docPool := pool.NewDocumentPool()
+	chunkPool := pool.NewChunkPool(2000)
 	mongoExtractor := newMongoExtractor(mockCollection, primitive.M{}, docPool, chunkPool)
 	var processedDocs []map[string]interface{}
 
@@ -121,8 +122,8 @@ func TestExtractor_Extract_SingleChunk(t *testing.T) {
 	mockCursor.On("Err").Return(nil)
 	mockCollection.On("Find", mock.Anything, mock.Anything, mock.Anything).Return(mockCursor, nil)
 
-	docPool := common.NewDocumentPool()
-	chunkPool := common.NewChunkPool(2000)
+	docPool := pool.NewDocumentPool()
+	chunkPool := pool.NewChunkPool(2000)
 	mongoExtractor := newMongoExtractor(mockCollection, primitive.M{}, docPool, chunkPool)
 	var processedDocs []map[string]interface{}
 
@@ -153,8 +154,8 @@ func TestExtractor_Extract_MultipleChunks(t *testing.T) {
 	mockCursor.On("Err").Return(nil)
 	mockCollection.On("Find", mock.Anything, mock.Anything, mock.Anything).Return(mockCursor, nil)
 
-	docPool := common.NewDocumentPool()
-	chunkPool := common.NewChunkPool(2000)
+	docPool := pool.NewDocumentPool()
+	chunkPool := pool.NewChunkPool(2000)
 	mongoExtractor := newMongoExtractor(mockCollection, primitive.M{}, docPool, chunkPool)
 	var processedDocs []map[string]interface{}
 	chunkCount := 0
@@ -177,8 +178,8 @@ func TestExtractor_Extract_FindError(t *testing.T) {
 	expectedErr := errors.New("find error")
 	mockCollection.On("Find", mock.Anything, mock.Anything, mock.Anything).Return(nil, expectedErr)
 
-	docPool := common.NewDocumentPool()
-	chunkPool := common.NewChunkPool(2000)
+	docPool := pool.NewDocumentPool()
+	chunkPool := pool.NewChunkPool(2000)
 	mongoExtractor := newMongoExtractor(mockCollection, primitive.M{}, docPool, chunkPool)
 	err := mongoExtractor.Extract(context.Background(), func(_ []map[string]interface{}) error {
 		return nil
@@ -201,8 +202,8 @@ func TestExtractor_Extract_DecodeError(t *testing.T) {
 	mockCursor.On("Close", mock.Anything).Return(nil)
 	mockColl.On("Find", mock.Anything, mock.Anything, mock.Anything).Return(mockCursor, nil)
 
-	docPool := common.NewDocumentPool()
-	chunkPool := common.NewChunkPool(2000)
+	docPool := pool.NewDocumentPool()
+	chunkPool := pool.NewChunkPool(2000)
 	mongoExtractor := newMongoExtractor(mockColl, primitive.M{}, docPool, chunkPool)
 
 	err := mongoExtractor.Extract(context.Background(), func(_ []map[string]interface{}) error {
@@ -230,8 +231,8 @@ func TestExtractor_Extract_CallbackError(t *testing.T) {
 	mockCursor.On("Close", mock.Anything).Return(nil)
 	mockCollection.On("Find", mock.Anything, mock.Anything, mock.Anything).Return(mockCursor, nil)
 
-	docPool := common.NewDocumentPool()
-	chunkPool := common.NewChunkPool(2000)
+	docPool := pool.NewDocumentPool()
+	chunkPool := pool.NewChunkPool(2000)
 	mongoExtractor := newMongoExtractor(mockCollection, primitive.M{}, docPool, chunkPool)
 	expectedErr := errors.New("callback error")
 
@@ -258,8 +259,8 @@ func TestExtractor_Extract_CursorError(t *testing.T) {
 	mockCursor.On("Err").Return(expectedErr)
 	mockCollection.On("Find", mock.Anything, mock.Anything, mock.Anything).Return(mockCursor, nil)
 
-	docPool := common.NewDocumentPool()
-	chunkPool := common.NewChunkPool(2000)
+	docPool := pool.NewDocumentPool()
+	chunkPool := pool.NewChunkPool(2000)
 	mongoExtractor := newMongoExtractor(mockCollection, primitive.M{}, docPool, chunkPool)
 	err := mongoExtractor.Extract(context.Background(), func(_ []map[string]interface{}) error {
 		return nil
@@ -290,8 +291,8 @@ func TestExtractor_Extract_WithFilter(t *testing.T) {
 	expectedFilter := primitive.M{"status": "active"}
 	mockCollection.On("Find", mock.Anything, expectedFilter, mock.Anything).Return(mockCursor, nil)
 
-	docPool := common.NewDocumentPool()
-	chunkPool := common.NewChunkPool(2000)
+	docPool := pool.NewDocumentPool()
+	chunkPool := pool.NewChunkPool(2000)
 	mongoExtractor := newMongoExtractor(mockCollection, expectedFilter, docPool, chunkPool)
 	var processedDocs []map[string]interface{}
