+++ internal/loader/loader_test.go
@@ -13,6 +13,7 @@ import (
 	"github.com/stretchr/testify/mock"
 
 	"mongo2dynamo/internal/common"
+	"mongo2dynamo/internal/pool"
 )
 
 const defaultMaxRetries = 5
@@ -56,8 +57,8 @@ func newTestLoader(client DBClient, table string, marshal MarshalFunc) *DynamoLo
 		table:      table,
 		marshal:    marshal,
 		maxRetries: defaultMaxRetries,
-		docPool:    common.NewDocumentPool(),
-		chunkPool:  common.NewChunkPool(1000),
+		docPool:    pool.NewDocumentPool(),
+		chunkPool:  pool.NewChunkPool(1000),
 	}
 }
