+++ internal/pool/pool.go
@@ -1,4 +1,4 @@
-package common
+package pool
 
 import (
 	"sync"
