+++ internal/pool/pool_test.go
@@ -1,4 +1,4 @@
-package common
+package pool
 
 import (
 	"fmt"
